Migrate submitting feedback to a job queue
Submitting feedback explored using a background IntentService to POST
to the API that a user wanted to submit feedback. This commit
introduces the Android Priority Job Queue library as a start at
introducing a common pattern for transmitting and receiving data that
is not tied so directly to the Activity lifecycle.

https://github.com/yigit/android-priority-jobqueue