Add mock API mode
Things were getting weird with Java 8 and Retrolambda and Dagger and
all the things. :boom: As a result, Java 8 support has been removed as
well in this commit. We’re back to Java 7, which is kind of sad, but
it’s solid and predictable. Now that the codebase is growing and the
debug mode is becoming more complex, this will remove an extra layer of
complexity from the build process.