Switch from Gson to Moshi
This commit switches the app’s JSON converter. Moshi is simpler and
promotes 1-to-1 mapping of JSON keys to Java field names. As such,
methods have been renamed to more properly match the JSON naming and
structure.

AutoValueMoshi generates JSON adapters for use by Moshi, which removes
a lot of JSON adapter boilerplate.

One trade off is that the custom Gson serializers that handled missing
data and would convert null or missing data into Null Objects have not
been carried over to Moshi. For now, views that access the Moshified
Java models need to check for null. This may be addressed in a future
commit.