Convert remaining jobs to use RxJava
This commit removes the job queue and event bus libraries in favor of
RxJava. The job queue / event bus architecture change didn’t really
work out, so the app is going back to an RxJava-based rest client,
currently tied to the UI layer. In the long run, it would be ideal to
separate the API data retrieval from the UI, but for now, this is
simple enough and will allow the app to work more predictably,
especially with error handling.