Update package name to org.selfconference (remove .android from it)
Why:
* For the 2019 self.conference Android release, we needed to release
the application under a different package name and store listing than
we had before, because we had lost the signing keys used for previous
releases, and therefore could not update the app any further under that
package name. We need to use a different package name for the
application going forward.

This change addresses the need by:
* Changing each package name of org.selfconference.android to
org.selfconference. I figured, since this is the org's only release on
Google Play, org.selfconference would suffice.